# docs/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "OUTLIER API",
  "version": "0.1.0"
 },
 "paths": {
...
  "/api/opportunities": {
   "get": {
    "tags": [
     "misc"
    ],
    "summary": "Opportunities",
    "operationId": "opportunities_api_opportunities_get",
    "parameters": [
     {
      "name": "tier",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Tier"
      }
     },
     {
      "name": "min_score",
      "in": "query",
      "required": false,
      "schema": {
       "type": "number",
       "default": 0,
       "title": "Min Score"
      }
     },
     {
      "name": "include_demo",
      "in": "query",
      "required": false,
      "schema": {
       "type": "boolean",
       "default": true,
       "title": "Include Demo"
      }
     },
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 1000,
       "default": 200,
       "title": "Limit"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/watchlist": {
   "get": {
    "tags": [
     "misc"
    ],
    "summary": "Watchlist",
    "operationId": "watchlist_api_watchlist_get",
    "parameters": [
     {
      "name": "include_archived",
      "in": "query",
      "required": false,
      "schema": {
       "type": "boolean",
       "default": false,
       "title": "Include Archived"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/watchlist/{listing_id}": {
   "put": {
    "tags": [
     "misc"
    ],
    "summary": "Upsert Watch",
    "operationId": "upsert_watch_api_watchlist__listing_id__put",
    "parameters": [
     {
      "name": "listing_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Listing Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/WatchIn"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "delete": {
    "tags": [
     "misc"
    ],
    "summary": "Remove Watch",
    "operationId": "remove_watch_api_watchlist__listing_id__delete",
    "parameters": [
     {
      "name": "listing_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Listing Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/listings/{listing_id}/feedback": {
   "post": {
    "tags": [
     "misc"
    ],
    "summary": "Add Feedback",
    "operationId": "add_feedback_api_listings__listing_id__feedback_post",
    "parameters": [
     {
      "name": "listing_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Listing Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/FeedbackIn"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/listings/{listing_id}/feedback/{feedback_id}": {
   "delete": {
    "tags": [
     "misc"
    ],
    "summary": "Delete Feedback",
    "operationId": "delete_feedback_api_listings__listing_id__feedback__feedback_id__delete",
    "parameters": [
     {
      "name": "listing_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Listing Id"
      }
     },
     {
      "name": "feedback_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Feedback Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/listings/{listing_id}/notes": {
   "post": {
    "tags": [
     "misc"
    ],
    "summary": "Add Note",
    "operationId": "add_note_api_listings__listing_id__notes_post",
    "parameters": [
     {
      "name": "listing_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Listing Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/NoteIn"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/listings/{listing_id}/notes/{note_id}": {
   "patch": {
    "tags": [
     "misc"
    ],
    "summary": "Patch Note",
    "operationId": "patch_note_api_listings__listing_id__notes__note_id__patch",
    "parameters": [
     {
      "name": "listing_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Listing Id"
      }
     },
     {
      "name": "note_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Note Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/NotePatch"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/listings/{listing_id}/outcome": {
   "put": {
    "tags": [
     "misc"
    ],
    "summary": "Upsert Outcome",
    "operationId": "upsert_outcome_api_listings__listing_id__outcome_put",
    "parameters": [
     {
      "name": "listing_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Listing Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/OutcomeIn"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/reference": {
   "get": {
    "tags": [
     "misc"
    ],
    "summary": "List Reference",
    "operationId": "list_reference_api_reference_get",
    "parameters": [
     {
      "name": "q",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Q"
      }
     },
     {
      "name": "domain",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Domain"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "post": {
    "tags": [
     "misc"
    ],
    "summary": "Create Reference",
    "operationId": "create_reference_api_reference_post",
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ReferenceIn"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/reference/{ref_id}": {
   "put": {
    "tags": [
     "misc"
    ],
    "summary": "Update Reference",
    "operationId": "update_reference_api_reference__ref_id__put",
    "parameters": [
     {
      "name": "ref_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Ref Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ReferenceIn"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "delete": {
    "tags": [
     "misc"
    ],
    "summary": "Delete Reference",
    "operationId": "delete_reference_api_reference__ref_id__delete",
    "parameters": [
     {
      "name": "ref_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Ref Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "FeedbackIn": {
    "properties": {
     "label": {
      "type": "string",
      "title": "Label"
     },
     "note": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Note"
     }
    },
    "type": "object",
    "required": [
     "label"
    ],
    "title": "FeedbackIn"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "NoteIn": {
    "properties": {
     "text": {
      "type": "string",
      "minLength": 1,
      "title": "Text"
     },
     "flagged": {
      "type": "boolean",
      "title": "Flagged",
      "default": false
     }
    },
    "type": "object",
    "required": [
     "text"
    ],
    "title": "NoteIn"
   },
...
   "WatchIn": {
    "properties": {
     "status": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Status"
     },
     "user_max_bid": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "User Max Bid"
     },
     "remind_minutes_before_end": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Remind Minutes Before End"
     },
     "notes": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Notes"
     },
     "archived": {
      "anyOf": [
       {
        "type": "boolean"
       },
       {
        "type": "null"
       }
      ],
      "title": "Archived"
     }
    },
    "type": "object",
    "title": "WatchIn"
   },
...
  }
 }
}